Designed title section. Added page route.
